tabadex_bot/requirements.txt
# requirements.txt

python-telegram-bot[ext]>=21.0.1
sqlalchemy>=2.0
psycopg2-binary>=2.9  # For PostgreSQL database
pydantic-settings>=2.0
aiohttp>=3.8 # For async API calls